requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
idna==3.6
pillow
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
django-nested-admin
